Use map to try to calculate current video address

+++ src/debugger/font.rs
@@ -10,7 +10,7 @@ use std::io::Cursor;
 
 pub struct Bitmap {
     // font: [u32; 8192]
-    pub font: Vec<u32>,
+    pub font: Vec<u8>,
 }
 impl Bitmap {
     pub fn new() -> Bitmap {
